Cleaned up some code related to the Summary authoring component. #2910
